requirements.txt
sounddevice>=0.4.1
wave
openai-whisper
numpy>=1.19.0
scikit-learn
transformers
torch>=1.9.0
scipy
soundfile>=0.10.3.post1
googletrans==3.1.0a0
speechbrain>=1.0.0
pyannote.audio
xgboost
librosa
langdetect
azure-cognitiveservices-speech
whisper 
numpy
pymongo
dnspython
python-dotenv
ffmpeg-python
azure-cognitiveservices-speech
pyaudio
streamlit
pygame
pydub